Testing, restructuring of backend
