+++ backend/src/__tests__/test.js
@@ -1,34 +1,74 @@
-const sFunctions = require('../serverFunctions.js');
-const makeid = require('../Room.js').makeid;
-const Room = require('../Room.js').Room;
-const User = require('../User.js').User;
+const dotenv = require('dotenv');
+const request = require('supertest');
+
+const lib = require('../lib');
+const makeid = require('../Classes/Room').makeid;
+const Room = require('../Classes/Room').Room;
+const User = require('../Classes/User').User;
+const server = require('../app').server;
+
+const users = require('../app').users;
+const rooms = require('../app').rooms;
+
+users.push(new User('', '', '', ''));
+users[0].id = 'User1';
+
+dotenv.config();
+const ipAddress = process.env.ADDRESS;
+const port = process.env.PORT;
+const portBack = process.env.PORTBACK;
+
+const uriBack = (ipAddress == 'localhost' ? 'http://' + ipAddress + ':' + portBack : 'https://' + ipAddress + ':' + port);
+const redirect_uri = uriBack + '/callback';
 
 describe('Basic Functions Tests', () => {
-	test('getRoomById', () => {
-		let rooms = [
-			{id: 'QUSDN'},
-			{id: 'UADBA'},
-			{id: 'DAJDN'},
-			{id: 'WNIFP'},
-			{id: 'YXNCB'}
-		];
-		expect(sFunctions.getRoomById('UADBA', rooms)).toBe(rooms[1]);
-		expect(sFunctions.getRoomById('WNIFP', rooms)).toBe(rooms[3]);
-		expect(sFunctions.getRoomById('YXNCB', rooms)).toBe(rooms[4]);
+	describe('getRoomById', () => {
+		test('The requested room is returned', () => {
+			let rooms = [
+				{id: 'QUSDN'},
+				{id: 'UADBA'},
+				{id: 'DAJDN'},
+				{id: 'WNIFP'},
+				{id: 'YXNCB'}
+			];
+			expect(lib.getRoomById('UADBA', rooms)).toBe(rooms[1]);
+			expect(lib.getRoomById('WNIFP', rooms)).toBe(rooms[3]);
+			expect(lib.getRoomById('YXNCB', rooms)).toBe(rooms[4]);
+		});
 	});
-	test('getUserById', () => {
-		let users = [
-			{id: 'blankesMichl'},
-			{id: 'Gabsii'},
-			{id: 'someoneElse'},
-			{id: 'AnotherOPne'},
-			{id: 'User'}
-		];
-		expect(sFunctions.getUserById('Gabsii', users)).toBe(users[1]);
+	
+	describe('getUserById', () => {
+		test('The requested User', () => {
+			let users = [
+				{id: 'blankesMichl'},
+				{id: 'Gabsii'},
+				{id: 'someoneElse'},
+				{id: 'AnotherOPne'},
+				{id: 'User'}
+			];
+			expect(lib.getUserById('Gabsii', users)).toBe(users[1]);
+			expect(lib.getUserById('blankesMichl', users)).toBe(users[0]);
+			expect(lib.getUserById('User', users)).toBe(users[4]);
+		});
 	});
-	test('makeId', () => {
-		let id = makeid(5);
-		expect(id).toHaveLength(5);
+
+	describe('makeId', () => {
+		test('If the id has the right length', () => {
+			let id = makeid(5);
+			expect(id).toHaveLength(5);
+		});
+		
+		test('If a unique id is returned every time', () => {
+			let ids = [];
+			for (let i = 0; i < 26; i++) {
+				let newId = makeid(1);
+				expect(ids.includes(newId)).toBe(false);
+				if (ids.includes(ids[i])) {
+					break;
+				}
+				ids.push(newId);
+			}
+		});
 	});
 });
 
@@ -100,5 +140,79 @@ describe('Room Functions Tests', () => {
 
 		let update = testRoom.getDifference(oldTestRoom);
 		expect(update.host).toEqual({name: 'Max Mustermann', voted: 'track1', img: 'https://upload.wikimedia.org/wikipedia/en/f/f9/Dat_Boi_%28resized_50%25%29.jpg'});
+	});
+});
+
+describe('App Tests', () => {
+	describe('Get / Method', () => {
+		test('It should responde with a little greeting', (done) =>  {
+			request(server).get('/').then((response) => {
+				expect(response.statusCode).toBe(200);
+				expect(response.text).toBe('Hello There');
+				done();
+			});
+		});
+	});
+
+	describe('Get /login Method', () => {
+		test('It should responde with a error if there is no referer', (done) =>  {
+			request(server).get('/login').then((response) => {
+				expect(response.statusCode).toBe(400);
+				expect(response.text).toBe('Login from the main page');
+				done();
+			});
+		});
+		test('It should responde with a redirect', (done) =>  {
+			request(server).get('/login').set({'Referer': 'http://jest.test/'}).then((response) => {
+				expect(response.statusCode).toBe(302);
+				let parameter = response.headers.location.split('&')[3];
+				let resRedirect = decodeURIComponent(parameter.split('=')[1]);
+				expect(resRedirect).toBe(redirect_uri);
+				done();
+			});
+		});
+	});
+
+	describe('Get /callback Method', () => {
+		test('It should responde with a error', (done) =>  {
+			request(server).get('/callback').then((response) => {
+				expect(response.statusCode).toBe(400);
+				done();
+			});
+		});
+	});
+
+	describe('Get /createRoom Method', () => {
+		test('It should responde with a error if no user is given', (done) =>  {
+			request(server).get('/createRoom').then((response) => {
+				expect(response.statusCode).toBe(400);
+				done();
+			});
+		});
+		test('It should responde with a redirect', (done) =>  {
+			request(server).get('/createRoom?id=User1').then((response) => {
+				expect(response.statusCode).toBe(302);
+				
+				expect(rooms.length > 0).toBeTruthy();
+				done();
+			});
+		});
+	});
+
+	describe('Get /createRoom Method', () => {
+		test('It should responde with a error if no user is given', (done) =>  {
+			request(server).get('/createRoom').then((response) => {
+				expect(response.statusCode).toBe(400);
+				done();
+			});
+		});
+		test('It should responde with a redirect', (done) =>  {
+			request(server).get('/createRoom?id=User1').then((response) => {
+				expect(response.statusCode).toBe(302);
+				
+				expect(rooms.length > 0).toBeTruthy();
+				done();
+			});
+		});
 	});
 });
