--- backend/src/constants.js
@@ -1,10 +1,0 @@
-let codes = {
-	SUCCESS: 200,
-	NOTFOUND: 404,
-	PLNOTFOUND: 414,
-	ERROR: 500
-};
-
-module.exports = {
-	codes: codes
-};
